+++ examples/minimal/boot.ts
@@ -1,7 +1,9 @@
 import express from 'express';
 import http from 'http';
-import { startPlatform, connectPlatformToExpress4, setAccessToken } from '@unchainedshop/platform';
-import { defaultModules, connectDefaultPluginsToExpress4 } from '@unchainedshop/plugins';
+import { startPlatform, setAccessToken } from '@unchainedshop/platform';
+// import { defaultModules, connectDefaultPluginsToExpress4 } from '@unchainedshop/plugins';
+// import { connect } from '@unchainedshop/api/express/index.js';
+
 import { log } from '@unchainedshop/logger';
 import seed from './seed.js';
 
@@ -10,30 +12,14 @@ const start = async () => {
   const httpServer = http.createServer(app);
 
   const engine = await startPlatform({
-    modules: defaultModules,
-    services: {},
-    options: {
-      payment: {
-        filterSupportedProviders: async ({ providers }) => {
-          return providers.sort((left, right) => {
-            if (left.adapterKey < right.adapterKey) {
-              return -1;
-            }
-            if (left.adapterKey > right.adapterKey) {
-              return 1;
-            }
-            return 0;
-          });
-        },
-      },
-    },
+    modules: {},
   });
 
   await seed(engine.unchainedAPI);
   await setAccessToken(engine.unchainedAPI, 'admin', 'secret');
 
-  connectPlatformToExpress4(app, engine);
-  connectDefaultPluginsToExpress4(app, engine);
+  // await connect(app, engine);
+  // connectDefaultPluginsToExpress4(app, engine);
 
   await httpServer.listen({ port: process.env.PORT || 3000 });
   log(`🚀 Server ready at http://localhost:${process.env.PORT || 3000}`);