+++ packages/api/src/api-index.ts
@@ -11,7 +11,6 @@ export * from './context.js';
 export * from './types.js';
 export * as acl from './acl.js';
 export * as errors from './errors.js';
-export * as express from './express/index.js';
 export * as roles from './roles/index.js';
 export { createContextResolver, getCurrentContextResolver, setCurrentContextResolver };